Refactor `RealmCourseProgress.insert` into `ProgressRepositoryImpl`
- Removed static `insert` method from `RealmCourseProgress.kt` to decouple database logic from the Realm model.
- Added a private `insertCourseProgress` method to `ProgressRepositoryImpl.kt`.
- Updated the bulk insert caller (`bulkInsertFromSync`) in `ProgressRepositoryImpl` to use the repository method.
